Terminal, Attack, and Art Templates
The terminal works and can be used by the putOnTerminal(String s)
method. It will never overflow and will be formatted correctly.

Attack now works for melee units, and will print to the terminal when an
attack lands.

I added some art that I quickly drew up. These need to be improved
because they suck.